Changed NautAngle to degrees, fixed operations
NautAngle now stores degrees internally and can be initialized with a float
value in degrees. self.distance_to(other) now is positive if travelling
east/north from self to other. __sub__ returns a float if difference is between
two NautAngles; it returns a NautAngle object if difference is between a
NautAngle and a float or int. Distance from -90 to +90 now comes out correctly
as 180.

windbreaker and poseidon with updated NautAngle